requirements.txt
# Core LLM + Transformers
torch>=2.1.0
transformers>=4.40.0
accelerate>=0.28.0
sentencepiece
huggingface_hub

# LangChain ecosystem (new modular structure)
langchain>=0.2.0
langchain-community>=0.2.0
langchain-core>=0.2.0

# For agents + tools
langchainhub
tiktoken

# Vector DB (for memory / retrieval)
faiss-cpu
chromadb

# Web search + scraping
duckduckgo-search
beautifulsoup4
requests
lxml

# Async support (important for multi-agent pipelines)
aiohttp
nest_asyncio

# Environment + utils
python-dotenv
pydantic>=2.0

# Optional (but highly useful)
rich          # better CLI logs
tqdm          # progress bars

# If using embeddings locally
sentence-transformers

# Optional (if you experiment with faster inference)
bitsandbytes

# for tavily
tavily-python

# streamlit for UI
streamlit